+++ client/src/pages/TownDetail/TownDetail.css
@@ -1,0 +1,5 @@
+.town-detail-main {
+  img {
+    width: 100%;
+  }
+}
